reorganize to separate api from app
