+++ dlx_rest/models.py
@@ -1,0 +1,6 @@
+from mongoengine import *
+
+## Setup some models for use
+
+class User(Document):
+    pass
